Update timeGen.h
modifié pour pouvoir afficher les millisecondes avec showOutput

+++ Projet-F1-V3/Projet OS/timeGen.h
@@ -5,5 +5,5 @@
 #ifndef PROJET_FORMULE_1_TIMEGEN_H
 #define PROJET_FORMULE_1_TIMEGEN_H
 
-int timeGenerator(int temps[]);
+int timeGenerator(float temps[]);
 #endif //PROJET_FORMULE_1_TIMEGEN_H
